[#64] move place storage
