+++ app/modules/places/components/list/place-list.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from "@angular/core";
 import { RouterExtensions } from 'nativescript-angular/router';
-import providePlaceStorageService from '../../../../shared/place/storage/provide-service';
-import { PlaceStorageService } from '../../../../shared/place/storage/place-storage.service';
+import providePlaceStorageService from '../../shared/storage/provide-service';
+import { PlaceStorageService } from '../../shared/storage/place-storage.service';
 import { PlaceImgPlaceholder } from '../../../../shared/place/place';
 
 const dialogs = require("ui/dialogs");
