+++ app/modules/places/components/shared/form-validation/place-form-validation.component.ts
@@ -5,8 +5,8 @@ import { ActivatedRoute } from '@angular/router';
 import { LoadingIndicator } from "nativescript-loading-indicator";
 
 import { PlaceSearchService } from '../../../../google-sdk/shared/place/search/place-search.service';
-import providePlaceStorageService from '../../../../../shared/place/storage/provide-service';
-import { PlaceStorageService } from '../../../../../shared/place/storage/place-storage.service';
+import providePlaceStorageService from '../../../shared/storage/provide-service';
+import { PlaceStorageService } from '../../../shared/storage/place-storage.service';
 import { CONTEXT_VALUES } from '../../../../../shared/place/place';
 
 @Component({
